refactor(site): extract docs search logic to src/lib/search.ts
Pulls tokenize, scoreEntry, rankResults and the SearchIndexEntry /
ScoredResult / SearchHeading types out of DocsSearchBar.tsx into a new
pure module at src/lib/search.ts. The component now imports from the
shared module; search-index.json.ts re-exports the entry types so any
existing importers keep working and the emitted JSON shape shares one
source of truth with the client-side scorer.

Scoring weights (W_TITLE, W_KEYWORD, W_HEADING, W_DESCRIPTION,
W_EXACT_PHRASE) are now exported so tests can reason about relative
weights symbolically instead of hardcoding numbers.

No behavior change. Verified with npm run build and npm run check
(0 errors).
